casino_bot/requirements.txt
# Casino Bot Dependencies
# Install with: pip install -r requirements.txt

python-telegram-bot>=20.0
